# Optimize InflaterHuffmanTree